# apps/web/public/openapi.json
{
	"openapi": "3.1.0",
	"info": {
		"title": "Ledger Agent Payments API",
		"version": "0.2.0",
		"description": "Agent-first API for submitting transaction intents. AI agents propose transactions; humans review and sign on Ledger hardware. Agents never touch private keys.\n\n## Quick Start\n1. Register your agent key (human signs authorization on Ledger)\n2. `POST /api/intents` with AgentAuth header to propose a transaction\n3. Share the returned `paymentUrl` with the human\n4. Poll `GET /api/intents/{id}` until status is `confirmed`, `rejected`, `failed`, or `expired`\n\n## Authentication\nAgents authenticate via the `Authorization: AgentAuth <timestamp>.<bodyHash>.<signature>` header.\n- `timestamp`: Unix epoch seconds (within 5 min of server time)\n- `bodyHash`: keccak256 of raw JSON body as hex (use `0x` for GET)\n- `signature`: EIP-191 personal_sign of `<timestamp>.<bodyHash>` with agent's secp256k1 key"
	},
	"servers": [
		{ "url": "https://agent-intents-web.vercel.app", "description": "Production" },
		{ "url": "/", "description": "Same-origin" }
	],
...
	"paths": {
...
		"/api/intents/status": {
			"post": {
				"tags": ["Intents"],
				"operationId": "updateIntentStatus",
				"summary": "Update intent status",
				"description": "Update the status of an intent. Agents (AgentAuth) can set: `executing`, `confirmed`, `failed`. Humans (session cookie) can set: `approved`, `rejected`, `authorized`, `broadcasting`. The caller must own the intent.\n\nThis is the preferred route (static path avoids Vercel dynamic-route issues).",
				"security": [{ "AgentAuth": [] }, { "SessionCookie": [] }],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": { "$ref": "#/components/schemas/UpdateIntentStatusByIdRequest" },
							"examples": {
								"agent_confirms": {
									"summary": "Agent confirms x402 settlement",
									"value": {
										"id": "int_1707048000_abc12345",
										"status": "confirmed",
										"note": "x402 payment settled successfully",
										"settlementReceipt": {
											"txHash": "0xabc123...",
											"network": "eip155:8453",
											"success": true
										}
									}
								},
								"agent_reports_failure": {
									"summary": "Agent reports failure",
									"value": {
										"id": "int_1707048000_abc12345",
										"status": "failed",
										"note": "x402 server returned 500 after payment"
									}
								}
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Status updated",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/IntentResponse" } }
						}
					},
					"400": {
						"description": "Invalid status or missing fields",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					},
					"401": {
						"description": "Authentication failed",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					},
					"403": {
						"description": "Not authorized (wrong owner or disallowed status transition)",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					},
					"404": {
						"description": "Intent not found",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					},
					"409": {
						"description": "Invalid state transition (e.g., confirmed → pending)",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					}
				}
			}
		},
...
	},
	"components": {
		"securitySchemes": {
			"AgentAuth": {
				"type": "apiKey",
				"in": "header",
				"name": "Authorization",
				"description": "Agent authentication. Format: `AgentAuth <timestamp>.<bodyHash>.<signature>`\n\n- `timestamp`: Unix epoch seconds (must be within 5 min of server time)\n- `bodyHash`: keccak256 of the raw JSON body as hex string (use `0x` for GET requests)\n- `signature`: EIP-191 personal_sign of the string `<timestamp>.<bodyHash>` using the agent's secp256k1 private key\n\nThe server recovers the signer address and matches it against registered agent public keys."
			},
			"SessionCookie": {
				"type": "apiKey",
				"in": "cookie",
				"name": "ai_session",
				"description": "Session cookie established via POST /api/auth/verify. Valid for 7 days. Used by the web UI for human actions."
			}
		},
		"schemas": {
...
			"ErrorResponse": {
				"type": "object",
				"required": ["success", "error"],
				"properties": {
					"success": { "type": "boolean", "const": false },
					"error": { "type": "string", "description": "Human-readable error message" }
				}
			},
			"IntentUrgency": {
				"type": "string",
				"enum": ["low", "normal", "high", "critical"],
				"default": "normal",
				"description": "Priority level. Affects recommended polling interval."
			},
			"IntentStatus": {
				"type": "string",
				"enum": [
					"pending",
					"approved",
					"rejected",
					"broadcasting",
					"authorized",
					"executing",
					"confirmed",
					"failed",
					"expired"
				],
				"description": "Intent lifecycle status. Terminal states: rejected, confirmed, failed, expired."
			},
			"PaymentCategory": {
				"type": "string",
				"enum": [
					"api_payment",
					"subscription",
					"purchase",
					"p2p_transfer",
					"defi",
					"bill_payment",
					"donation",
					"other"
				],
				"description": "Category of payment (for display and analytics)"
			},
			"X402Resource": {
				"type": "object",
				"properties": {
					"url": { "type": "string", "format": "uri", "description": "The URL being paid for" },
					"description": { "type": "string" },
					"mimeType": { "type": "string" }
				},
				"required": ["url"]
			},
			"X402AcceptedExactEvm": {
				"type": "object",
				"description": "x402 payment requirement for EVM exact scheme (EIP-3009 TransferWithAuthorization)",
				"required": ["network", "amount", "asset", "payTo"],
				"properties": {
					"scheme": { "type": "string", "const": "exact" },
					"network": { "type": "string", "description": "CAIP-2 network ID (e.g., eip155:8453)" },
					"amount": {
						"type": "string",
						"description": "Amount in atomic units (e.g., 10000 = 0.01 USDC)"
					},
					"asset": { "type": "string", "description": "ERC-20 token contract address" },
					"payTo": { "type": "string", "description": "Recipient address" },
					"maxTimeoutSeconds": { "type": "integer" },
					"extra": {
						"type": "object",
						"properties": {
							"name": { "type": "string", "description": "EIP-712 domain name (e.g., USD Coin)" },
							"version": { "type": "string", "description": "EIP-712 domain version (e.g., 2)" },
							"decimals": { "type": "integer" }
						}
					}
				}
			},
			"X402Context": {
				"type": "object",
				"description": "Full x402 payment context stored on an intent",
				"properties": {
					"resource": { "$ref": "#/components/schemas/X402Resource" },
					"accepted": { "$ref": "#/components/schemas/X402AcceptedExactEvm" },
					"authorization": {
						"type": "object",
						"description": "EIP-3009 TransferWithAuthorization parameters (populated after signing)",
						"properties": {
							"from": { "type": "string" },
							"to": { "type": "string" },
							"value": { "type": "string" },
							"validAfter": { "type": "string" },
							"validBefore": { "type": "string" },
							"nonce": { "type": "string" }
						}
					},
					"signature": {
						"type": "string",
						"description": "EIP-712 signature of the authorization (populated after signing)"
					},
					"paymentSignatureHeader": {
						"type": "string",
						"description": "Base64-encoded PAYMENT header value the agent should send to the x402 server. Only returned to the owning agent."
					},
					"paymentPayload": {
						"type": "object",
						"description": "Structured payment payload (only returned to owning agent)"
					},
					"settlementReceipt": {
						"type": "object",
						"description": "Receipt from x402 server after settlement",
						"properties": {
							"txHash": { "type": "string" },
							"network": { "type": "string" },
							"amount": { "type": "string" },
							"settledAt": { "type": "string" },
							"blockNumber": { "type": "integer" },
							"success": { "type": "boolean" },
							"error": { "type": "string" }
						}
					},
					"expiresAt": {
						"type": "string",
						"format": "date-time",
						"description": "When the x402 authorization expires"
					}
				}
			},
			"TransferIntent": {
				"type": "object",
				"required": ["type", "token", "amount", "recipient", "chainId"],
				"properties": {
					"type": { "type": "string", "const": "transfer" },
					"token": { "type": "string", "description": "Token symbol (e.g., USDC, ETH)" },
					"tokenAddress": {
						"type": "string",
						"description": "ERC-20 contract address (auto-filled for known tokens)"
					},
					"tokenLogo": { "type": "string", "description": "URL to token logo image" },
					"amount": {
						"type": "string",
						"description": "Human-readable amount (e.g., '50', '0.01')"
					},
					"amountWei": { "type": "string", "description": "Amount in wei for precision" },
					"recipient": { "type": "string", "description": "Destination address (0x...)" },
					"recipientEns": { "type": "string", "description": "ENS name if resolved" },
					"chainId": {
						"type": "integer",
						"description": "Chain ID: 8453 (Base), 11155111 (Sepolia), 84532 (Base Sepolia)"
					},
					"memo": { "type": "string", "description": "Human-readable reason for the transaction" },
					"resource": {
						"type": "string",
						"description": "x402: URL of the resource being paid for"
					},
					"merchant": {
						"type": "object",
						"description": "Merchant/payee information",
						"properties": {
							"name": { "type": "string" },
							"url": { "type": "string" },
							"logo": { "type": "string" },
							"verified": { "type": "boolean" }
						}
					},
					"category": { "$ref": "#/components/schemas/PaymentCategory" },
					"x402": { "$ref": "#/components/schemas/X402Context" }
				}
			},
			"Intent": {
				"type": "object",
				"required": [
					"id",
					"userId",
					"agentId",
					"agentName",
					"details",
					"urgency",
					"status",
					"createdAt",
					"statusHistory"
				],
				"properties": {
					"id": {
						"type": "string",
						"description": "Unique intent ID (format: int_<timestamp>_<uuid>)"
					},
					"userId": { "type": "string", "description": "Wallet address of the human signer" },
					"agentId": { "type": "string", "description": "Agent identifier" },
					"agentName": { "type": "string", "description": "Agent display name" },
					"details": { "$ref": "#/components/schemas/TransferIntent" },
					"urgency": { "$ref": "#/components/schemas/IntentUrgency" },
					"status": { "$ref": "#/components/schemas/IntentStatus" },
					"trustChainId": {
						"type": "string",
						"description": "Trustchain identity (wallet address) — set when created by an authenticated agent"
					},
					"createdByMemberId": {
						"type": "string",
						"format": "uuid",
						"description": "Agent member UUID"
					},
					"createdAt": { "type": "string", "format": "date-time" },
					"expiresAt": { "type": "string", "format": "date-time" },
					"reviewedAt": { "type": "string", "format": "date-time" },
					"broadcastAt": { "type": "string", "format": "date-time" },
					"confirmedAt": { "type": "string", "format": "date-time" },
					"txHash": {
						"type": "string",
						"description": "Transaction hash (populated after confirmation)"
					},
					"txUrl": { "type": "string", "description": "Block explorer URL" },
					"statusHistory": {
						"type": "array",
						"description": "Ordered audit trail of status changes",
						"items": {
							"type": "object",
							"required": ["status", "timestamp"],
							"properties": {
								"status": { "$ref": "#/components/schemas/IntentStatus" },
								"timestamp": { "type": "string", "format": "date-time" },
								"note": { "type": "string" }
							}
						}
					}
				}
			},
			"IntentResponse": {
				"type": "object",
				"required": ["success", "intent"],
				"properties": {
					"success": { "type": "boolean", "const": true },
					"intent": { "$ref": "#/components/schemas/Intent" }
				}
			},
...
			"UpdateIntentStatusByIdRequest": {
				"type": "object",
				"required": ["id", "status"],
				"description": "Used with POST /api/intents/status (preferred). Includes intent ID in the body.",
				"properties": {
					"id": { "type": "string", "description": "Intent ID (e.g., int_1707048000_abc12345)" },
					"status": { "$ref": "#/components/schemas/IntentStatus" },
					"txHash": { "type": "string", "description": "Transaction hash" },
					"note": { "type": "string", "description": "Audit note" },
					"paymentSignatureHeader": {
						"type": "string",
						"description": "x402: base64-encoded payment signature header"
					},
					"paymentPayload": {
						"type": "object",
						"additionalProperties": true,
						"description": "x402: structured payment payload"
					},
					"settlementReceipt": {
						"type": "object",
						"additionalProperties": true,
						"description": "x402: settlement receipt from server"
					},
					"expiresAt": {
						"type": "string",
						"format": "date-time",
						"description": "Override expiration timestamp"
					}
				}
			},
...
		}
	}
}